requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-apscheduler==0.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.3
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
yapf==0.43.0
yfinance==0.2.50